+++ CHANGES.md
@@ -1,5 +1,19 @@
 # Calendar.js - Change Log:
 
+### Version 1.8.6:
+- Fixed the testing HTML for the minimized version of Calendar.js.
+- Renamed the following CSS classes (sorry, this will cause some style breakages):
+  - "checkboxContainer" is now "checkbox-container".
+  - "textHeader" is now "text-header".
+  - "radioButtonsContainer" is now "radio-buttons-container".
+  - "radioButtonContainer" is now "radio-button-container".
+  - "radioButton" is now "radio-button".
+- Improvements to the documentation and added support to show which parameters are option for public functions.
+- Fixed a fault that caused the expand button in the main display (in the cells) to remain visible when dragging an event.
+- Added a new custom event "onEventClick", which is triggered when an event is clicked (passes the event details).
+- Added a new custom event "onEventDoubleClick", which is triggered when an event is double clicked (passes the event details).  This event is only fired if event editing is disabled.
+- Added improved documentation for custom triggers (explains what can be passed, when they are called, etc).
+
 ### Version 1.8.5:
 - Added a custom CSS class for the text headers used in the Advanced area of the Search dialog.
 - Added a new custom event "onEventDragStart", which is triggered when dragging an event is started (the event details are passed to the function).
@@ -16,7 +30,7 @@
 - The public function addHolidays() will now only accept a valid array of holidays.
 - The public function removeHolidays() will now only accept a valid array of string holiday names.
 - Added "locked" property support for events, which states if an event can be edited (it can still be removed).
-- Added disabled text colour support for input fields.
+- Added disabled text color support for input fields.
 - Improved the disabled CSS that is applied to check boxes/radio buttons that are disabled.
 
 ### Version 1.8.3: